// src/app/Helper/models/WishlistModel.ts
export class WishlistModel {
  ApproveDate?: Date;
  ApproveDateString: string;
  ApprovedBy: string;
  ApprovedByID?: number;
  Barcode: string;
  CreatedAt?: Date;
  CreatedAtString: string;
  CreatedBy: string;
  CreatedByID?: number;
  ID: number;
  IsActive?: boolean;
  IsApproved?: boolean;
  IsPurchased?: boolean;
  Outlet: string;
  OutletID?: number;
  ProductVariant: string;
  ProductVariantID?: number;
  PurchaseDate?: Date;
  PurchaseDateString: string;
  PurchasedBy: string;
  PurchasedByID?: number;
  Quantity?: number;
  Remarks: string;
  UpdatedAt?: Date;
}